requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
vertexai==0.0.1
pillow==10.2.0
google-cloud-aiplatform==1.43.0
langchain==0.1.10
langchain-google-genai==0.0.9
langchain-google-vertexai==0.1.0